Start a subprocess in which to do op commands ...
We start the subprocess the first time we need op. It waits until
it is informed we have op. If we can't get it fast enough, we give up
and end the process.

After having op, we start processing messages that should be performed
as an operator. While we keep getting messages, we hand them off to the
outbound message queue. Once we haven't had a new op command for a while,
we take away our ops and end the subprocess.

The rationale for this

- we wait to perform the op action until we are sure we have op. This
  gets rid of that ugly sleep that ate up an outbound message queue token
- we also wait to deop ourselves until we are reasonably sure we aren't
  going to need it again for a while. This saves yet another token in
  the short term.

So if we need to go on a banning spree, we can ban 100 people with only
102 messages (op, 100 bans, deop) instead of 300 (op, ban, deop, op, ban,
deop, op ...) (assuming no wasted token on calling sleep)

Also in this commit

- misc. comment improvements