Fix race condition on startup.
ii needs to be one of the first things to start. It also needs to create
the FIFOs it uses quickly. If it doesn't, then when other parts of the
bot try to write to the FIFOs, they end up creating regular files
automatically creating regular files. Then when ii reads these regular
files, it floods the IRC server.